requirements.txt
torch==2.9.1+cpu
torchvision==0.24.1+cpu
fastapi==0.121.3
uvicorn==0.38.0
pillow==12.0.0
python-multipart==0.0.20
requests==2.32.5
httpx==0.25.0 
--extra-index-url https://download.pytorch.org/whl/cpu